adding changes and resetting workspace
