+++ ExperienceRight-BackCapTS/Controllers/UserProfileController.cs
@@ -6,17 +6,19 @@ using ExperienceRight_BackCapTS.Repositories;
 
 namespace ExperienceRight_BackCapTS.Controllers
 {
-    //[Authorize]
+    [Authorize]
     [Route("api/[controller]")]
     [ApiController]
     public class UserProfileController : ControllerBase
     {
         private readonly IUserProfileRepository _userProfileRepository;
         private readonly IUserTypeRepository _userTypeRepository;
-        public UserProfileController(IUserProfileRepository userProfileRepository, IUserTypeRepository userTypeRepository)
+        private readonly IBusinessRepository _businessRepository;
+        public UserProfileController(IUserProfileRepository userProfileRepository, IUserTypeRepository userTypeRepository, IBusinessRepository businessRepository)
         {
             _userProfileRepository = userProfileRepository;
             _userTypeRepository = userTypeRepository;
+            _businessRepository = businessRepository;
         }
 
         [HttpGet]
@@ -25,17 +27,36 @@ namespace ExperienceRight_BackCapTS.Controllers
             return Ok(_userProfileRepository.GetAllUsersANDBusinessz());
         }
 
+
+
+
+
+
+        //[HttpGet("{firebaseUserId}")]
+        //public IActionResult GetByFirebaseUserId(string firebaseUserId)
+        //{
+        //    var userProfile = _userProfileRepository.GetUserByFirebaseUserId(firebaseUserId);
+        //    if (userProfile == null)
+        //    {
+        //        return NotFound();
+        //    }
+        //    return Ok(userProfile);
+        //}
+
+
+
+
         [HttpGet("customer/{firebaseUserId}")]
         public IActionResult GetUserProfile(string firebaseUserId)
         {
             return Ok(_userProfileRepository.GetUserByFirebaseUserId(firebaseUserId));
         }
 
-        [HttpGet("business/{firebaseUserId}")]
-        public IActionResult GetBusinessProfile(string firebaseUserId)
-        {
-            return Ok(_userProfileRepository.GetBusinessByFirebaseUserId(firebaseUserId));
-        }
+        //[HttpGet("business/{firebaseUserId}")]
+        //public IActionResult GetBusinessProfile(string firebaseUserId)
+        //{
+        //    return Ok(_userProfileRepository.GetBusinessByFirebaseUserId(firebaseUserId));
+        //}
 
         [HttpGet("user/{id}")]
         public IActionResult GetUserProfileById(int id)
@@ -43,12 +64,24 @@ namespace ExperienceRight_BackCapTS.Controllers
             return Ok(_userProfileRepository.GetProfileById(id));
         }
 
-        [HttpPost("UserProfile")]
-        public IActionResult UserProfilePost(UserProfile userProfile)
+        //[HttpPost]
+        //public IActionResult UserProfile(UserProfile userProfile)
+        //{
+        //    userProfile.CreateDateTime = DateTime.Now;
+        //    userProfile.UserTypeId = UserType.Anonymous_ID;
+        //    _userProfileRepository.AddUserProfile(userProfile);
+        //    return CreatedAtAction(
+        //        nameof(GetUserProfile),
+        //       // nameof(GetByFirebaseUserId),
+        //        new { firebaseUserId = userProfile.FirebaseUserId },
+        //        userProfile);
+        //}
+
+        [HttpPost]
+        public IActionResult Post(UserProfile userProfile)
         {
             userProfile.CreateDateTime = DateTime.Now;
             userProfile.UserTypeId = UserType.Anonymous_ID;
-            //userProfile.IsActive = true;
             _userProfileRepository.AddUserProfile(userProfile);
             return CreatedAtAction(
                 nameof(GetUserProfile),
@@ -56,18 +89,30 @@ namespace ExperienceRight_BackCapTS.Controllers
                 userProfile);
         }
 
-        [HttpPost("BusinessProfile")]
-        public IActionResult BusinessProfilePost(UserProfile userProfile)
-        {
-            userProfile.CreateDateTime = DateTime.Now;
-            userProfile.UserTypeId = UserType.Business_ID;
-            //userProfile.IsActive = true;
-            _userProfileRepository.AddBusinessProfile(userProfile);
-            return CreatedAtAction(
-                nameof(GetUserProfile),
-                new { firebaseUserId = userProfile.FirebaseUserId },
-                userProfile);
-        }
+
+
+
+
+
+
+        //[HttpPost("BusinessProfile")]
+        //public IActionResult BusinessProfilePost(int id, UserProfile userProfile, Business business )
+        //{
+
+        //    //id = userProfile.Id;
+        //    userProfile.CreateDateTime = DateTime.Now;
+        //    userProfile.UserTypeId = UserType.Business_ID;
+        //    //userProfile.IsActive = true;
+        //    _userProfileRepository.AddBusinessProfile(userProfile, id);
+        //    //_userProfileRepository.GetUserProfile(id);
+        //    //_businessRepository.GetUserBusinessById(userProfileId);
+        //    _businessRepository.AddBusiness(business);
+        //    return CreatedAtAction(
+        //        nameof(GetUserProfile),
+        //        new { firebaseUserId = userProfile.FirebaseUserId },
+        //        userProfile);
+
+        //}
 
         //[HttpPut("{id}")]
         //public ActionResult Put(UserProfile userProfile)
