+++ SQLQuery1.sql
@@ -110,3 +110,14 @@
                           FROM UserProfile up
                                LEFT JOIN UserType ut on up.UserTypeId = ut.Id
                          WHERE ut.Id= 2 
+
+
+                                                 SELECT up.Id AS UserProfileId, up.FirebaseUserId, up.FirstName, up.LastName, up.DisplayName, 
+                               up.Email, up.CreateDateTime, up.ProfileImageLocation, up.UserTypeId, 
+                               ut.Name AS UserTypeName,
+                               b.Id AS BusinessId, b.EstablishmentName, b.Bio, b.Address, b.HoursOfOperation, b.Phone, b.UserProfileId AS BusinessUserProfileId, b.CategoryId,
+                               c.Id, c.Name AS CategoryName
+                          FROM UserProfile up
+                               LEFT JOIN UserType ut on up.UserTypeId = ut.Id
+                               LEFT JOIN Business b on b.UserProfileId = up.Id
+                               LEFT JOIN Category c on b.CategoryId = c.Id
